+++ language/compiler/src/check/tests/expressions/assignment/parameter.rs
@@ -17,7 +17,7 @@ function bump(value: int32): int32 {
         r#"
 === annotated ===
 function bump(value: int32): int32 {
-    value = 2 as int32;
+    value = 2;
     return value;
 }
 
@@ -27,10 +27,10 @@ function bump(value: int32): int32 {
 /// @type.symbol symbol=value source="value: int32" type=int32
 
     value = 2;
-    /// @type.node source="value = 2" type=2
+    /// @type.node source="value = 2" type=int32
     /// @type.node source=value type=int32
     /// @resolution.name source=value target=value
-    /// @type.node source=2 type=2
+    /// @type.node source=2 type=int32
 
     return value;
     /// @type.node source=value type=int32
@@ -97,7 +97,7 @@ function bump(value: int32): int32 {
         r#"
 === annotated ===
 function bump(value: int32): int32 {
-    value += 2 as int32;
+    value += 2;
     return value;
 }
 
@@ -110,7 +110,7 @@ function bump(value: int32): int32 {
     /// @type.node source="value += 2" type=int32
     /// @type.node source=value type=int32
     /// @resolution.name source=value target=value
-    /// @type.node source=2 type=2
+    /// @type.node source=2 type=int32
     /// @resolution.call source="value += 2" parameters=() return=int32 kind=builtin builtin=binary.add
 
     return value;
